Fix bug related to screen drawing in SDL
Source of the bug was that even if the global instace of the ScreenManager
owned the SDL window, SDL is prepared to work with the window in the
thread that created the window. I considered two possible solutions: controlling
(showing) the window from the main thread (outside the states) or deleting
the window each time a new RdScreen is set. I selected the second one. A
minor flaw derived from this choice is that the window is repositioned each
time I delete it.