Cleanup included headers
The presence or absence of #include's indicates whether a dependency
should be treated as 'public' or 'private'. This commit aims at removing
unused declarations and including the missing ones (even when no
compilation errors occurred). The choice for their location is relevant:
elements that correspond to the public interface should be located in
the header file. On the other hand, private dependencies must not polute
any .hpp and reside in the source file (.cpp) instead.

See also http://stackoverflow.com/a/26038443

The following convention was adopted:

* Header files should only hold the #include's used within the .hpp
* Source files contain any *additional* #include's used only in the .cpp
* Order of #include directives:

  #include <(system header 1)>
  #include <(system header 2)>
  
  #include <(dependency header 1, e.g. ZBar/SDL/YARP)>
  #include <(dependency header 2, e.g. ZBar/SDL/YARP)>
  
  #include "(RD header 1)"
  #include "(RD header 2)"
  #include "(Macros.hpp/Utils.hpp)"