Force CMAKE_INSTALL_PREFIX as rel path to itself
'install(... DESTINATION ${CMAKE_INSTALL_PREFIX})' crashes on the
'make package' step because of interpreting the destination as an
absolute path (even if it's the same path as our install prefix).
Relevant error reports are invoked from cmake_install.cmake script.
An empty string does not work, a blank space creates the directory
'/usr/local/ /' on UNIX both on install and package.